[TODO] 완료: WritePage에서 WriteElement 초까지 관리
